image_agent/requirements.txt
crawl4ai>=0.4.0
pymupdf>=1.24.0
pymupdf4llm>=0.0.10
google-generativeai>=0.8.0
httpx>=0.27.0
click>=8.1.0
pyyaml>=6.0.0
aiofiles>=24.1.0
pillow>=10.0.0
python-dotenv>=1.0.0
rich>=13.0.0
pydantic>=2.0.0
